internal/dag: move validation struct lookups
Build processors for both the ExtensionService and HTTPProxy types
have to deal with `UpstreamValidation`. Move the helpers that look up
`UpstreamValidation` and `DownstreamValidation` types to the Kubernetes
cache so that they are accessible to all the build processors.

This updates #2713.
