+++ internal/dag/cache.go
@@ -14,6 +14,7 @@
 package dag
 
 import (
+	"errors"
 	"fmt"
 	"strings"
 	"sync"
@@ -438,6 +439,43 @@ func (kc *KubernetesCache) LookupSecret(name types.NamespacedName, validate func
 	return s, nil
 }
 
+func (kc *KubernetesCache) LookupUpstreamValidation(uv *projectcontour.UpstreamValidation, namespace string) (*PeerValidationContext, error) {
+	if uv == nil {
+		// no upstream validation requested, nothing to do
+		return nil, nil
+	}
+
+	secretName := types.NamespacedName{Name: uv.CACertificate, Namespace: namespace}
+	cacert, err := kc.LookupSecret(secretName, validCA)
+	if err != nil {
+		// UpstreamValidation is requested, but cert is missing or not configured
+		return nil, fmt.Errorf("invalid CA Secret %q: %s", secretName, err)
+	}
+
+	if uv.SubjectName == "" {
+		// UpstreamValidation is requested, but SAN is not provided
+		return nil, errors.New("missing subject alternative name")
+	}
+
+	return &PeerValidationContext{
+		CACertificate: cacert,
+		SubjectName:   uv.SubjectName,
+	}, nil
+}
+
+func (kc *KubernetesCache) LookupDownstreamValidation(vc *projectcontour.DownstreamValidation, namespace string) (*PeerValidationContext, error) {
+	secretName := types.NamespacedName{Name: vc.CACertificate, Namespace: namespace}
+	cacert, err := kc.LookupSecret(secretName, validCA)
+	if err != nil {
+		// PeerValidationContext is requested, but cert is missing or not configured.
+		return nil, fmt.Errorf("invalid CA Secret %q: %s", secretName, err)
+	}
+
+	return &PeerValidationContext{
+		CACertificate: cacert,
+	}, nil
+}
+
 // DelegationPermitted returns true if the referenced secret has been delegated
 // to the namespace where the ingress object is located.
 func (kc *KubernetesCache) DelegationPermitted(secret types.NamespacedName, targetNamespace string) bool {
@@ -471,4 +509,12 @@ func (kc *KubernetesCache) DelegationPermitted(secret types.NamespacedName, targ
 		}
 	}
 	return false
+}
+
+func validCA(s *v1.Secret) error {
+	if len(s.Data[CACertificateKey]) == 0 {
+		return fmt.Errorf("empty %q key", CACertificateKey)
+	}
+
+	return nil
 }
